make specs pass with Rails 5
this PR may upgrade this gem to work with Rails 5. it
certainly upgrades all the specs to pass under Rails
5, which is not the same thing, but obviously it's a
decent proxy. it also removes `Gemfile.lock` since
[`Gemfile.lock` should not be in version control for a
gem](http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/).

this PR also removes the implementation of a spec which depended
on behavior that is no longer possible, namely instantiating an
ActiveRecord model for which no table exists. the spec wasn't really a
spec about doing that anyway, so its expectation remains in the spec
file, but it's been made pending.